+++ stdlib/tests/crypto/falcon.rs
@@ -15,8 +15,8 @@ use test_utils::{
     },
     expect_exec_error_matches,
     proptest::proptest,
-    rand::{rand_value, rand_vector},
-    FieldElement, QuadFelt, Word, WORD_SIZE,
+    rand::rand_vector,
+    Word,
 };
 use vm_core::StarkField;
 
@@ -26,38 +26,18 @@ const Q: u64 = (M - 1) / 2;
 const N: usize = 512;
 const J: u64 = (N * M as usize * M as usize) as u64;
 
-#[test]
-fn test_set_to_zero() {
-    let source = "
+const PROBABILISTIC_PRODUCT_SOURCE: &str = "
     use.std::crypto::dsa::rpo_falcon512
 
     begin
-        # write bytes in the first and last addresses of the region to be zeroed
-        push.1.2.3.4 mem_storew.1000 dropw
-        push.1.2.3.4 mem_storew.3044 dropw
+        #=> [PK, ...]
+        push.0
+        #=> [h_ptr, PK, ...]
 
-        # This address should be untouched
-        push.1.2.3.4 mem_storew.3048 dropw
-
-        push.1000
-        exec.rpo_falcon512::set_to_zero
-
-        # Assert that output pointer is 1000 + 4 * 512 = 3048
-        push.3048 assert_eq
+        exec.rpo_falcon512::load_h_s2_and_product
+        #=> [...]
     end
     ";
-
-    let expected_memory = {
-        let mut memory = vec![0_u64; N * 4];
-        // addresses [3048, 3052) (not zeroed)
-        memory.extend_from_slice(&[1, 2, 3, 4]);
-
-        memory
-    };
-
-    let test = build_test!(source, &[]);
-    test.expect_stack_and_memory(&[], 1000_u32, &expected_memory);
-}
 
 #[test]
 fn test_falcon512_norm_sq() {
@@ -80,7 +60,7 @@ fn test_falcon512_norm_sq() {
 }
 
 #[test]
-fn test_falcon512_diff_mod_q() {
+fn test_falcon512_diff_mod_m() {
     let source = "
     use.std::crypto::dsa::rpo_falcon512
 
@@ -151,111 +131,34 @@ proptest! {
 }
 
 #[test]
-fn test_falcon512_powers_of_tau() {
-    let source = "
-    use.std::crypto::dsa::rpo_falcon512
-
-    begin
-        exec.rpo_falcon512::powers_of_tau
-    end
-    ";
-
-    // Compute powers of a quadratic field element from 0 to 512.
-    let tau = rand_value::<QuadFelt>();
-    let tau_ptr = 0_u32;
-    let (tau_0, tau_1) = ext_element_to_ints(tau);
-
-    let expected_memory = powers_of_tau(tau);
-    let stack_init = [tau_ptr.into(), tau_0, tau_1];
-
-    let test = build_test!(source, &stack_init);
-    let expected_stack = &[u64::from(tau_ptr) + (N as u64 + 1_u64) * 4];
-    test.expect_stack_and_memory(expected_stack, tau_ptr, &expected_memory);
-}
-
-const PROBABILISTIC_PRODUCT_SOURCE: &str = "
-    use.std::crypto::dsa::rpo_falcon512
-
-    begin
-        #=> [PK, ...]
-        mem_load.0
-        #=> [h_ptr, PK, ...]
-
-        exec.rpo_falcon512::load_h_s2_and_product
-        #=> [tau1, tau0, tau_ptr, ...]
-
-        exec.rpo_falcon512::powers_of_tau
-        #=> [zeros_ptr, ...]
-
-        exec.rpo_falcon512::set_to_zero
-        #=> [c_ptr, ...]
-
-        drop
-        #=> [...]
-
-        push.2048    # tau_ptr
-        push.4100   # zeroes_ptr (tau_ptr + 4 * 513)
-        push.0      # h ptr
-
-        #=> [h_ptr, zeros_ptr, tau_ptr, ...]
-
-        exec.rpo_falcon512::probabilistic_product
-    end
-    ";
-
-#[test]
 fn test_falcon512_probabilistic_product() {
-    // Create two random polynomials and multiply them.
-    let h = Polynomial::new(random_coefficients());
-    let s2 = Polynomial::new(random_coefficients());
+    // create two random polynomials and generate the input operand stack and advice stack to
+    // the probabilistic product test procedure
+    let h: Polynomial<Felt> = Polynomial::new(random_coefficients());
+    let s2: Polynomial<Felt> = Polynomial::new(random_coefficients());
+    let (operand_stack, advice_stack): (Vec<u64>, Vec<u64>) =
+        generate_data_probabilistic_product_test(h, s2, false);
 
-    let pi = mul_modulo_p(h.clone(), s2.clone());
-
-    // Lay the polynomials in the advice stack, h then s2 then pi = h * s2.
-    let mut h_array = to_elements(h.clone());
-    h_array.extend(to_elements(s2.clone()));
-    h_array.extend(pi.iter().map(|a| Felt::new(*a)));
-    let advice_stack: Vec<u64> = h_array.iter().map(|&e| e.into()).collect();
-
-    // Compute hash of h and place it on the stack.
-    let binding = Rpo256::hash_elements(&to_elements(h.clone()));
-    let h_hash = binding.as_elements();
-    let h_hash_copy: Vec<u64> = h_hash.iter().map(|felt| (*felt).into()).collect();
-    let stack_init = vec![h_hash_copy[0], h_hash_copy[1], h_hash_copy[2], h_hash_copy[3]];
-
-    let test = build_test!(PROBABILISTIC_PRODUCT_SOURCE, &stack_init, &advice_stack);
+    let test = build_test!(PROBABILISTIC_PRODUCT_SOURCE, &operand_stack, &advice_stack);
     let expected_stack = &[];
     test.expect_stack(expected_stack);
 }
 
-#[ignore = "needs horner_eval_base op"]
 #[test]
 fn test_falcon512_probabilistic_product_failure() {
-    // Create a polynomial pi that is not equal to h * s2.
+    // create two random polynomials and generate the input operand stack and advice stack to
+    // the probabilistic product test procedure
     let h: Polynomial<Felt> = Polynomial::new(random_coefficients());
     let s2: Polynomial<Felt> = Polynomial::new(random_coefficients());
-    let h_wrong: Polynomial<Felt> = Polynomial::new(random_coefficients());
+    let (operand_stack, advice_stack): (Vec<u64>, Vec<u64>) =
+        generate_data_probabilistic_product_test(h, s2, true);
 
-    let pi = mul_modulo_p(h_wrong.clone(), s2.clone());
-
-    // Lay the polynomials in the advice stack, h then s2 then pi = h_wrong * s2.
-    let mut h_array = to_elements(h.clone());
-    h_array.extend(to_elements(s2.clone()));
-    h_array.extend(pi.iter().map(|a| Felt::new(*a)));
-    let advice_stack: Vec<u64> = h_array.iter().map(|&e| e.into()).collect();
-
-    // Compute hash of h and place it on the stack.
-    let binding = Rpo256::hash_elements(&to_elements(h.clone()));
-    let h_hash = binding.as_elements();
-    let h_hash_copy: Vec<u64> = h_hash.iter().map(|felt| (*felt).into()).collect();
-
-    let stack_init = vec![h_hash_copy[0], h_hash_copy[1], h_hash_copy[2], h_hash_copy[3]];
-    let test = build_test!(PROBABILISTIC_PRODUCT_SOURCE, &stack_init, &advice_stack);
+    let test = build_test!(PROBABILISTIC_PRODUCT_SOURCE, &operand_stack, &advice_stack);
 
     expect_exec_error_matches!(
         test,
         ExecutionError::FailedAssertion{ clk, err_code, err_msg }
-        if clk == RowIndex::from(18843) && err_code == 0 && err_msg.is_none()
+        if clk == RowIndex::from(3182) && err_code == 0 && err_msg.is_none()
     );
 }
 
@@ -333,42 +236,8 @@ fn generate_test(
 
 // HELPER FUNCTIONS
 // ================================================================================================
-// Helper function to convert a quadratic extension field element into a tuple of elements in the
-// underlying base field and convert them into integers.
-fn ext_element_to_ints(ext_elem: QuadFelt) -> (u64, u64) {
-    let base_elements = ext_elem.to_base_elements();
-    (base_elements[0].as_int(), base_elements[1].as_int())
-}
 
-/*
-    For an element `tau := (tau0, tau1)` in the quadratic extension field, computes all its powers
-    `tau^i` for `i = 0,..., 512` and stores them in a vector of length 2048 (word size * 512).  The
-    first two field elements of the ith word are the elements of tau^i, and the second two field
-    elements are the previous power of tau, tau^(i - 1).  Used to test powers of tau procedure.
-    Example:
-    [1, 0, 0, 0, tau_0, tau_1, 1, 0, (tau^2)_0, (tau^2)_1, tau_0, tau_1, (tau^3)_0, (tau^3)_1,
-    (tau^2)_0, (tau^2)_1, ...]
-*/
-
-fn powers_of_tau(tau: QuadFelt) -> Vec<u64> {
-    let mut tau_power: QuadFelt;
-    let mut elem_0: u64;
-    let mut elem_1: u64;
-    let mut expected_memory = vec![0; (N + 1) * WORD_SIZE];
-    expected_memory[0] = 1;
-
-    for i in 1..N + 1 {
-        tau_power = tau.exp(i as u64);
-        (elem_0, elem_1) = ext_element_to_ints(tau_power);
-        expected_memory[i * WORD_SIZE] = elem_0;
-        expected_memory[i * WORD_SIZE + 1] = elem_1;
-        expected_memory[i * WORD_SIZE + 2] = expected_memory[i * WORD_SIZE - WORD_SIZE];
-        expected_memory[i * WORD_SIZE + 3] = expected_memory[i * WORD_SIZE - 3];
-    }
-    expected_memory
-}
-
-// Create random coefficients in the range of a polynomial in M.
+/// Creates random coefficients of a polynomial in the range (0..M).
 fn random_coefficients() -> Vec<Felt> {
     let mut res = Vec::new();
     for _i in 0..N {
@@ -377,13 +246,14 @@ fn random_coefficients() -> Vec<Felt> {
     res
 }
 
-/* Multiplies two polynomials over Z_p\[x\] without reducing modulo p. Given that the degrees
-of the input polynomials are less than 512 and their coefficients are less than the modulus
-q equal to M = 12289, the resulting product polynomial is guaranteed to have coefficients less
-than the Miden prime.
-Note that this multiplication is not over Z_p\[x\]/(phi).
-*/
-pub fn mul_modulo_p(a: Polynomial<Felt>, b: Polynomial<Felt>) -> [u64; 1024] {
+/// Multiplies two polynomials over Z_p\[x\] without reducing modulo p.
+///
+/// Given that the degrees of the input polynomials are less than 512 and their coefficients are
+/// less than the modulus M = 12289, the resulting product polynomial is guaranteed to have
+/// coefficients less than the Miden prime.
+///
+/// Note that this multiplication is not over Z_p\[x\]/(phi).
+fn mul_modulo_p(a: Polynomial<Felt>, b: Polynomial<Felt>) -> [u64; 1024] {
     let mut c = [0; 2 * N];
     for i in 0..N {
         for j in 0..N {
@@ -393,6 +263,42 @@ pub fn mul_modulo_p(a: Polynomial<Felt>, b: Polynomial<Felt>) -> [u64; 1024] {
     c
 }
 
-pub fn to_elements(poly: Polynomial<Felt>) -> Vec<Felt> {
+/// Returns the coefficients of a polynomial.
+fn to_elements(poly: Polynomial<Felt>) -> Vec<Felt> {
     poly.coefficients.to_vec()
+}
+
+/// Generates the data needed to execute the probabilistic product test.
+fn generate_data_probabilistic_product_test(
+    h: Polynomial<Felt>,
+    s2: Polynomial<Felt>,
+    test_failure: bool,
+) -> (Vec<u64>, Vec<u64>) {
+    let pi = mul_modulo_p(h.clone(), s2.clone());
+
+    // lay the polynomials in order h then s2 then pi = h * s2
+    let mut polynomials = if test_failure {
+        to_elements(Polynomial::new(random_coefficients()))
+    } else {
+        to_elements(h.clone())
+    };
+    polynomials.extend(to_elements(s2.clone()));
+    polynomials.extend(pi.iter().map(|a| Felt::new(*a)));
+
+    // get the challenge point and push it to the advice stack
+    let digest_polynomials = Rpo256::hash_elements(&polynomials);
+    let challenge = (digest_polynomials[0], digest_polynomials[1]);
+    let mut advice_stack = vec![challenge.0.as_int(), challenge.1.as_int()];
+
+    // push the polynomials to the advice stack
+    let polynomials: Vec<u64> = polynomials.iter().map(|&e| e.into()).collect();
+    advice_stack.extend_from_slice(&polynomials);
+
+    // compute hash of h and place it on the stack.
+    let binding = Rpo256::hash_elements(&to_elements(h.clone()));
+    let h_hash = binding.as_elements();
+    let h_hash_copy: Vec<u64> = h_hash.iter().map(|felt| (*felt).into()).collect();
+    let operand_stack = vec![h_hash_copy[0], h_hash_copy[1], h_hash_copy[2], h_hash_copy[3]];
+
+    (operand_stack, advice_stack)
 }